upstream release
Upstream is now fully migrated to python3 and python2 is no longer supported or
compatible. Good riddance! We now move back to a single package which replaces
all previous alternatives.

Moves from python-sip (4.x) to sip5 and sip-builder to build.

Also note one test is being ignored since it crashes in Qt, but only in
makechrootpkg/nspawn, and works fine on the host.